tests/integration/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
httpx[http2]>=0.27.0
psutil>=5.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
zstandard>=0.22.0
brotli>=1.1.0
websockets>=15.0  # WebSocket support for uvicorn backend
websocket-client>=1.9.0  # WebSocket client for integration tests
grpcio>=1.60.0  # gRPC Python library
grpcio-tools>=1.60.0  # Protocol buffer compiler for gRPC
